resolved todos
